\name{CVTrainIndexGene}
\alias{CVTrainIndexGene}
\title{ CV index matrix }
\description{
 Get index matrix for the training data in cross-validation.
 Only for internal use. Not available to users.
}
\usage{
CVTrainIndexGene(data, fold, seed)
}
\arguments{
  \item{data}{ the data frame generated by \code{expData.R} }
  \item{fold}{ the fold of cross-validation }
  \item{seed}{ random seed }
}
\value{
 A matrix of \code{TRUE} or \code{FALSE}. \code{TRUE} for training data, and \code{FALSE} for testing data. Each row strands for one fold.
}
\keyword{ models }
