\name{getNullCount}
\alias{getNullCount}
\title{ counts under null model }
\description{
 Get the counts of reads for each position under the null hypothesis
 (all reads have the same sequencing preferences).
}
\usage{
getNullCount(data)
}
\arguments{
  \item{data}{ the data frame generated by \code{expData.R} }
}
\value{
 a numeric vector stores the counts
}
\examples{
 # read and expand the data
 data(g1_part) # for real data, please use read.csv, like g1 <- read.csv("g1.csv")
 data <- expData(g1_part, 2, 3)

 null.count <- getNullCount(data)
 null.dev <- getDev(null.count, data$count)
}
\keyword{ models }
