\name{getOriLogPref}
\alias{getOriLogPref}
\title{ initialize for mart }
\description{
 Get the original log preferences, that is, the logarithms of gene 
 expression levels. Here the gene with zero count will be replaced by
 a small count. Only for internal use. Not available to users.
}
\usage{
getOriLogPref(data, small_count = 0.5)
}
\arguments{
  \item{data}{ the data frame generated by \code{expData.R} }
  \item{small_count}{ the small count that will replace the zero count }
}
\value{
 a numeric vector of the log preferences
}
\keyword{ models }
