\name{getWeight}
\alias{getWeight}
\title{ weight vector }
\description{
 Get the weight vector for MART. Only for internal use. Not available to users.
}
\usage{
getWeight(data)
}
\arguments{
  \item{data}{ the data frame generated by \code{expData.R} }
}
\value{
 a numeric vector of the weights
}
\keyword{ models }
