\name{setOriOffset}
\alias{setOriOffset}
\title{ initialize }
\description{
 Get the original offsets for Poisson linear model, that is, the 
 logarithms of gene expression levels. Only for internal use. Not available to users.
}
\usage{
setOriOffset(data)
}
\arguments{
  \item{data}{ the data frame generated by \code{expData.R} }
}
\value{
 a numeric vector of the offsets
}
\keyword{ models }
