\name{updateOffset}
\alias{updateOffset}
\title{ update offsets }
\description{
 Update the offsets for Poisson linear model, that is, the logarithms of
 gene expression levels when we take sequencing preferences into account.
 Only for internal use. Not available to users.
}
\usage{
updateOffset(data, pred_count, log_expr)
}
\arguments{
  \item{data}{ the data frame generated by \code{expData.R} }
  \item{pred_count}{ the counts fitted by Poisson linear model }
  \item{log_expr}{ The log expression levels (offsets) we had before this iteration. }
}
\value{
 a numeric vector of the offsets
}
\keyword{ models }
